Moving cnv frequencies functions to their own file
